feat: component 구조 변경
