--- front-end/src/components/Button/styles.js
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const searchButton = styled.img`
-  width: 50px;
-  height: 50px;
-  cursor: pointer;
-  margin: 25px;
-`;
