+++ front-end/src/components/Dashboard/Bottom/styles.js
@@ -1,0 +1,1 @@
+import styled from "styled-components";
