--- front-end/src/components/DashboardBottom/styles.js
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-export const btnContainer = styled.div`
-  width: 100%;
-  height: fit-content;
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  position: absolute;
-  align-items: end;
-  justify-items: end;
-  margin-top: 25%;
-`;
